xNAV upgrades for Navcoin Core 7.0.0 (#862)
* use nbits for signaling

* optimize xnav verification

* fix candidate serialization

* new output ser encoding

* confidential tokens

* serialize newSer

* confidential tokens only need their own H gen

* ser fixes

* token db

* rpc commands, vdata predicates

* final commit ctokens

* fix include

* fix

* add program_actions

* burn

* do not rw newser

* RandomShuffle

* remove newser

* include <algorithm>

* fix duplicated

* read token from program

* attach extra data in mint and burn programs

* nfts

* fixes for nfts

* fix segfault

* fix

* fix generator

* do not cache token balances

* TokenId class

* encrypted trading msg exchange

* use shared_ptr for range proof

* fix ser

* fix compressor

* do not combine token burn as fee

* outcompressor ser

* abandon xnav conflict

* fix sendnft

* allow nftid to be 0

* fix sendnft

* tx fixes

* correctly abandon

* isempty

* fix total supply

* Token stressor (#69)

* add stressor for tokens/nfts

* update description

* add xNAV transactions

* turn off gdb

* remove dao functions

* show xnav balance

* fix select private coins

* use spendingkey to determine blsct output

* use mcl for generator tokens

* fix nft send (#70)

* add blsct also from mempool

* update

* more fixes

* depth 0 is only spent if in pool

* fix transactionkeys rpc

+++ src/versionbits.h
@@ -133,7 +133,7 @@ struct VersionBitsCache
 };
 
 ThresholdState VersionBitsState(const CBlockIndex* pindexPrev, const Consensus::Params& params, Consensus::DeploymentPos pos, VersionBitsCache& cache);
-uint32_t VersionBitsMask(const Consensus::Params& params, Consensus::DeploymentPos pos);
+uint64_t VersionBitsMask(const Consensus::Params& params, Consensus::DeploymentPos pos);
 bool IsVersionBitRejected(const Consensus::Params& params, Consensus::DeploymentPos pos);
